Rename module to just `route_map`
It's not clear that we need the module to include `_rs`
