+++ poetry_build.py
@@ -8,7 +8,7 @@ def build(setup_kwargs: Dict[str, Any]) -> None:
     Add rust_extensions to the setup dict
     """
     setup_kwargs["rust_extensions"] = [
-        RustExtension("starlite.route_map_rs", path="extensions/rust/Cargo.toml", binding=Binding.PyO3)
+        RustExtension("starlite.route_map", path="extensions/rust/Cargo.toml", binding=Binding.PyO3)
     ]
     setup_kwargs["zip_safe"] = False
